colocate drag and drop logic
